Kind of fixed
So, I forgot what did I type yesterday. Basically, when an user does clic on the close button on a "site wide alert" it will crete a session with the alert on the server, therefore, when the user navigates around the site, he/she won't see the alert again until he closes the browser and comes back.